long->double

+++ navit/android/src/org/navitproject/navit/NavitDownloadSelectMapActivity.java
@@ -198,7 +198,7 @@ public class NavitDownloadSelectMapActivity extends ExpandableListActivity {
         if (mapIndex != null) {
             int mi = Integer.parseInt(mapIndex);
             // check if download larger than 4 GiB
-            long maxsz = Math.pow(2, 32);
+            double maxsz = Math.pow(2, 32);
             if (Double.parseDouble(android.os.Build.VERSION.RELEASE.replaceAll("(\\d+[.]\\d+)(.*)","$1")) >= 8) {
                // check if download larger than 16 GiB on Nougat and up
                 maxsz = Math.pow(2, 34);
